Moved DB code into store module.
